Fixed context error in ImageTexture
ImageTexture was using finalize to have the GC delete the OpenGL texture
when the object was cleaned up, but this happens on a seperate thread,
and as such doesn't have the correct context. Changed name of method so
it will have to me called manually from the main thread to be invoked.